itzmenu-extractor/pyproject.toml
[build-system]
requires = [ 'setuptools>=69.1.1' ]
build-backend = 'setuptools.build_meta'

[project]
name = 'itzmenu_extractor'
dynamic = ['version']
readme = 'README.md'
license = { text = 'Apache License 2.0' }
authors = [
    { name = 'Louke', email = 'git@jannsen.org' }
]
classifiers = [
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: Apache License 2.0',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: Implementation :: CPython',
]
dependencies = [
    'beanie>=1.25.0',
    'motor>=3.4.0',
    'typing_extensions>=4.9.0',
    'pydantic_settings>=2.2.1',
    'requests>=2.31.0',
    'pillow>=10.2.0',
    'img2table[gcp]>=1.2.11',
    'pytesseract>=0.3.10',
    'apscheduler>=3.10.4',
    'holidays>=0.45'
]

[project.optional-dependencies]
test = [
    'pytest>=8.0.2',
    'pytest_httpserver>=1.0.10'
]


[tool.setuptools.package-data]
'*' = [ '*.json', '*.jpg' ]

[tool.setuptools.dynamic]
version = { attr = 'itzmenu_extractor.__version__' }